Resize Doge appropriately. Aspect ratio constant.
In the updated replaceTwitterLogo function, a new imgWrapper div is created to wrap the image. The imgWrapper div has a paddingBottom value that is calculated based on the aspect ratio of the image to ensure that the image maintains its aspect ratio. The img tag is then positioned absolutely within the imgWrapper div with width and height set to 100% and object-fit set to contain to ensure that the entire image is visible even if it needs to be scaled down.

The dogeContainer div is set to have a position of relative so that the imgWrapper div can be positioned absolutely within it.

With these modifications, the dogeCrop.png image should now resize proportionally when the nested divs get too thin to display the entire image while maintaining its aspect ratio.